move reflection tests to own file
